requirements.txt
absl-py==0.12.0
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
cmake==3.18.4.post1
cycler==0.10.0
docx2txt==0.8
EbookLib==0.17.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.4.4
flatbuffers==1.12
gast==0.3.3
h5py==2.10.0
imutils==0.5.4
joblib==1.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==1.1.1
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
pandas==1.2.3
pdf2image==1.14.0
pdfminer.six==20181108
Pillow==8.1.2
pkgconfig==1.5.2
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.1
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2021.1
PyYAML==5.4.1
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.23.1
scipy==1.6.1
six==1.12.0
sortedcontainers==2.3.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
SQLAlchemy==1.3.23
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
textract==1.6.3
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
tzlocal==1.5.1
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.3.7


